net: use if-watch and allow dual-stack sockets
When binding to INADDR(6)_ANY, use the `if-watch` crate to dynamically
discover and update the listen addresses. This simplifies building on
windows, as we don't need winpcap anymore. It also allows to obtain the
current listen addresses without handling IO errors.

When binding to an ipv6 address, allow the socket to operate in
dual-stack mode if the platform supports it.

Closes #532
Signed-off-by: Kim Altintop <kim@monadic.xyz>