tls: add some warn logging when certificate resolution fails
This is motivated by a fun behaviour of the docker-compose setup, which
puzzled me for a moment: the dynamically-scaled peers will all bind to
the same port. Because 127.0.0.1 is routable inside the docker network,
connection attempts will end up on a random(?) peer, which may or may
not match the peer id used.

Amusingly, everything still works, but the only log output is a
propagated error from rustls, which is a bit confusing.

Signed-off-by: Kim Altintop <kim@monadic.xyz>