replication: Ensure git stream is closed
Make sure for all incoming git operations that both recveiving and
sending streams are driven to completion so all QUIC control frames make
it to the peer. This is crucial so that counters of the internal state
machine of quinn are increased correctly, which in turn allows new
streams to be opened.

The newly introduced smoke test should fail if the `RecvStream` in the
git server is not polled.

Fixes #602

Signed-off-by: Alexander Simmerl <a.simmerl@gmail.com>