daemon: run all async tests on the multi-threaded runtime
In conjunction with #650, we see async tests intermittently hanging
indefinitely when the per-test runtime shuts down. This may have to do
with some blocking operations in `daemon` not being properly
`spawn_blocking`-ed (and thus blocking the single runtime thread), or
some mysterious interaction with threads spawned from within the async
program.

In any case, using the multi-threaded runtime appears to fix it.

Signed-off-by: Kim Altintop <kim@monadic.xyz>