seed: Improve request fulfilment
This change-set brings two improvements:

Firstly it caps the time a call on the `NodeHandle` can take by
wrapping the recv on the response channel to 3 seconds. During seed
operation a common issue that has been observed is that these requests
would pile up if something in the pipes is clogged. With this change the
downstream frontend implementation and generally all consumers can
bubble up errors more elegantly by timing out or presenting proper
protocol native responses like a 504 in the case of an HTTP server.

Additionally the branch in the seed event loop is not crashing anymore,
but instead will log the error.

Signed-off-by: Alexander Simmerl <a.simmerl@gmail.com>