seed: Never block event loop
If one of the arms in the select of the main event loop block for
contention on a backing reource simply slow interactions with remote
peers the entire seed can become unresponsive. With the help of upper
bounds on internal queues and actual timeouts on requests of all sorts,
it should be viable to spawn the subroutines as they stream in.

Signed-off-by: Alexander Simmerl <a.simmerl@gmail.com>