librad: executor shim
Passing around a wrapper around the async runtime allows us to
approximate scoped teardown of spawned tasks, commit on a defined
cancellation semantics (drop aborts), centralise instrumentation, and
potentially allow becoming runtime-agnostic in the future.

This patch also fixed teardown of temporary directories in the
integration tests by creating and dropping the runtime explicitly (as
opposed to the #[tokio::test] macro). This is due to the ill-defined
cancellation semantics of tokio, which offers no other way to ensure all
tasks are dropped at a defined point in time.

Signed-off-by: Kim Altintop <kim@monadic.xyz>