daemon: parameterise Peer
We parameterise the Peer that we accept in `daemon::state` with a
parameter S. This parameter must implement Clone + Signer, which mimics
the interface of librad. This allows us to accept BoxedSigner,
SecretKey, etc.

This becomes useful when we want to use a Peer<SecretKey> from the
testnet code.

We also add a `with_peer` constructor for the daemon Peer type so that
we construct one with an existing librad Peer.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>