net: fix CBOR decoder behaviour on incomplete input
The `futures_codec::Decoder` impl was adapted from the original
`serde`-based CBOR decoder to `minicbor`. It turns out that this ported
over a bug: when the input buffer does not contain enough data, its
offset must **not** be advanced. Otherwise, the next poll will attempt
to decode from that offset instead of the start of the object, which
will obviously fail.

Fixes #648
Signed-off-by: Kim Altintop <kim@eagain.st>