storage::pool: split config into read and write
We extract the signer and fetchers for a pool config into a `Write`
struct. The `Config` type now has a `W` parameter representing whether
it can create read-only storage, or if it can create read-write storage.

We add type synonyms for ReadConfig and ReadWriteConfig, as well as
adding the `Manager` impl for ReadConfig.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>