profile: refactor Profile
While scoping out the introduction of creating, getting, and setting
profiles, I saw some room for refactoring to better see the flow of
profile creation and loading.

This patch introduces the `profile::id` module which contains the
`ProfileId` type. It surfaces the many ways to construct a ProfileId:
* From a UUID
* Read from RAD_PROFILE
* Read from active_profile
* Read from active_profile or default to creating one

Before this change, a profile ID was simply a String, but now that it is
behind a newtype wrapper, it's more obvious that it must go through
validation during construction.

The other introduction to this change is the RadHome enum which captures
where the Profile is living: XDG or some given Root.

The construction of a Profile is now computed in terms of RadHome and
ProfileId.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>