daemon: add shutdown signal
`Peer::run` now properly cleans up the network stack when run to
completion. To this end it accepts a shutdown signal as an argument.
This seems to fix deadlocks that occur when the daemon is torn down
multiple times [1].

Instead of copying the approach from `seed/src/lib.rs` we use a shutdown
future and port this approach to the seed. The reason for this is that
the seed also did not handle shutdown properly. If a termination signal
was sent it did not break the main loop. Moreover we did not react to
shutdown requests while awaiting `peer.bind()`.

[1]: https://github.com/radicle-dev/radicle-upstream/issues/2133

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>