rad-identities: introduction of rad-identities package
The majority of the work is porting over code from `copy` and modifying
it to be generic over the type of identity being used.

The package is split into top-level modules:
* any
* local
* project
* person
* rad_refs
* refs
* tracking

For each of those there's a corresponding module under the cli module
which performs all the setup logic followed by calling the top-level
function, mentioned above.

The cli module culminates in the args and main modules, where the former
defines all the arguments requried for calling the evaluation functions,
and the latter calling those eval functions based on the given command.

The field module captures lens-like Has traits so that we can write
functionality that can be re-used across both Person and Project. These
traits allow us to talk about the fields on those two identities
polymorphically.

The git hierarchy provides utilities for setting up working copies.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>