librad: fix deadlock when waiting for idle
We fix a deadlock when we wait for the quic endpoint to become idle.

The deadlock happens when we wait for `endpoint.wait_idle()` while we
are simultaneously still trying to connect to a remote that does not
respond (i.e. it neither accepts nor refuses the connection) in
`connect` called by `librad::net::protocol::io::discovered()`.

To make `endpoint.wait_idle()` future finish we need to advance the
connection task so we drop it after the future has finished instead of
before.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>