net: retain outgoing uni streams
While using unidirectional streams for gossip/membership RPCs simplified
a lot of state management, using a stream per message was a less than
ideal choice: streams are cheap but not free, and applying flow control
in that model is hard to impossible.

In order to tie those streams to the connection lifetime, and not
further complicate connection tracking, we introduce a notion of
"reservation" of a fixed number of uni streams which can be borrowed
mutably to send RPCs. Necessarily, the streams are behind mutexes,
which should be fine for the use case.

Signed-off-by: Kim Altintop <kim@eagain.st>