git-protocol: make stateless
While working on replication v3, it becomes apparent that layering more
state handling on top of a stateful transport is getting quite
convoluted. Thus make all git protocol v2 interactions stateless, ie. a
"delegate" represents a single command.

Note that this requires to assemble the capabilities advertisement
ourselves, as `git upload-pack` cannot be taught to send it when using
`--stateless-rpc`.

Signed-off-by: Kim Altintop <kim@eagain.st>