link-replication: version 3 of replication, sans-io
Third attempt to get replication right, as a new crate. Assumes git
protocol v2 support.

This is still a work-in-progress, as testability did not pan out yet the
way it should. As it is not wired up yet, it appears preferable to apply
improvements incrementally.

Note that no IO impls backing the effect traits are provided yet.

Signed-off-by: Kim Altintop <kim@eagain.st>