daemon: qualify announcement path
The previous implementation used `join` over `into_qualified`, this
would result in the following in a reference of the form `rad/id`. This
reference can be ambiguous -- either meaning `refs/rad/id` or
`refs/heads/rad/id`.

We know, however, that this should be `refs/rad/id`. So using the
`into_qualified` function would give us the correct path of
`refs/rad/id`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>