daemon: remove sync
With rere and eventually grafting, we will not need the sync mechanism.

Currently we have been deploying seeds with rere disabled, but this
makes upgrading less maintainable as we need to keep track of an extra
branch. We did this because of some emergent effects between rere and
sync, so this also allows us to move forward with rere.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>