rad-exe: top-level rad CLI
This is the first draft of the top-level rad CLI library -- to become the
`rad` executable.

In its current form, it defines the global arguments and performs the
sanitisation of those arguments, as per the RFC. It has mock `profile`
subcommand that does nothing. It allows for an external `rad-<command>`
to be called spawning a subprocess to execute it.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>