git: ignore error due to concurrent sigrefs update
When two `Refs::update` race, only one of them can succeed, because the
ref update only succeeds if the expected parent commit is what the ref
currently points to (ref updates are atomic). While there is a chance
that the winning update commits a slightly outdated view of the refs
tree, this is tolerable, because every modification of the namespace
will trigger a sigrefs update, and thus the state converges.

Not so great, however, is to propagate the commit failure due to a
modified parent as an error, as this may interrupt the replication flow,
leaving the namespace in a potentially half-arsed state.

Therefore, teach `Refs::update` to ignore this kind of error, and return
the most recent committed state instead.

Fixes #761
Signed-off-by: Kim Altintop <kim@eagain.st>