remote: remove fetch and push refspecs before delete
If a remote is deleted while containing refspecs it will delete the
references that match those refspecs. This is not the desired behaviour
in `with_tmp_copy`. There's no `git2` method for removing an entire
section, so instead we remove the `fetch` and `push` entries. When we
delete the remote, the entire entry is removed, and our fetched
references are preserved.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>