link-identities: introduce Ext type
The way to extend a payload was only possible if the type lived as a
Rust type, since HasNamespace's Url value must be statically known. This
bars an extension being passed in from the outside, for example as a CLI
input.

This patch introduces an Ext type that consists of a value and a Url. A
type that implements HasNamespace can be converted into an Ext. The
`set_ext` and `with_ext` functions are now defined in terms of Into<Ext>.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>