node: Add library to wire up a running p2p node
First iterative implementation of RFC 0696 focusing on driving the core
peer protocol and establishing the harness for configuration of a running node.
Rite of passage for it will be replacement of the ephemeral peer in the e2e
networkss, as that requires a well behaved peer with a set of knobs exposed.

Structurally the majority of the implementation lives in a library crate
node-lib. The code is rather Mario-esque as it is primarily plumbing of code
that either existed in other crates backing binaries or newish code doing the
same thing for core pieces.

There are still large pieces missing, which a tracked in #722 and will be
piled on-top as patches to keep this already big delta focused. The declared
initial goal is to get linkd into a state where it can run as bootstrap/seed
node.

Signed-off-by: Alexander Simmerl <a.simmerl@gmail.com>