linkd: Add binary
A very thin and light wrapper around the `node::run` exposed from the
node lib crate, so `linkd` exists as a deploy target.

Bootstrapping a whole new workspace which is excluded from root, allows
for a checked in `Cargo.lock` without affecting the lib crates.
Historically this used to be the role of radicle-bins.

Signed-off-by: Alexander Simmerl <a.simmerl@gmail.com>