e2e: Replace ephemeral peer with linkd
Now that a binary exists which drives the core protocol and has all the
knobs exposed that the e2e network tests want to turn there is no need
anymore for the ad-hocish epehmeral peer implementation.

This concludes the first chapter of linkd's development as the declared
goal of gaining some confidence through at the integration layer for it
succeeded.

Signed-off-by: Alexander Simmerl <a.simmerl@gmail.com>