node: Implement tracking
Adds a new subroutine to the node library which takes care of tracking
projects based on a pre-defined set of either everything, certain Urns,
anything from a certain peers, or mix of the latter two. This is a port
of the tracking code that currently exists in the seed lib and will be
exposed as configuration in a follow commit.

Signed-off-by: xla <sefl@xla.is>