rad-clib: extend ssh keys functionality
radicle-keystore added list and remove functions to the ssh-agent
functionality. rad-clib can utilise these for functions for testing and
removing key material from the agent.

signer is extended to check if the key is present before attempting to
get the signing key.

remove_signer is self explanatory.

is_signer_present lists the keys and checks if the profile's key is in that
list.

sign gets the signer from the ssh-agent and signs a payload.

verify verifies a signature and payload match for the profile's key.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>