rad-profile: extend ssh commands & fix ssh add
We extend the ssh commands for rad-profile to include removing a key,
checking if a key is present on the ssh-agent, signing with a key, and
verifying a signature and payload with a key.

Since there are multiple ssh commands, we split this into a subcommand.

Due to an issue in thrussh (see
https://nest.pijul.com/pijul/thrussh/discussions/52), constraints were
not being handled properly. This is fixed in the fork we are using[0].
We add a constraint in the person creation test to convince ourselves
that this is indeed working.

[0]: https://github.com/FintanH/thrussh/tree/generic-agent

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>