rad, node, bins: configurable ssh auth socket
We would like to be able to configure which socket the ssh-agent listens
on when running tests.

First, the dependency on radicle-keystore is bumped in link-crypto to be able
to do this.

This required rad-clib's ssh related functions to accept an SshAgent as
an argument for adding, removing, and listing keys. To make this
configurable, rad-clib receives a new enum called SshAuthSock. It can
either be Env or Uds -- where the latter contains a file path. This type
can change the default SshAgent's path -- using SSH_AUTH_SOCK -- to the
defined one, if Uds is given.

This had a cascading effect in threading the SshAuthSock through all
rad-identities, rad-profile, and node-lib functions that relied on the
ssh functionality.

As rad-identities and rad-profile but rely on this value, it is
introduced as a new global argument in rad-exe's Global args, with the
default being Env.

This all allowed us to define a test utility for spinning up an
`ssh-agent` using a temporary socket and running the ssh functions in the
test in that bracketed function `with_ssh_agent`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>