rad-identities: add Display struct
When using the rad-identities CLI, I realised that there was no way to
actually get a Urn for using it as a parameter in other commands.

To rectify this, we add a Display struct, generic over the payload. The
Display struct consists of a Urn and the respective payload. This allows
us to output the urn and the payload together on stdout. There are
helper traits for converting various identities into a Display, as well
a fmt::Display implementation -- for completeness sake.

In the person, project, and any modules we use type synonyms to refine
the generic to their specific payloads.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>