rad-identities: create project before initialising working copy
The initialisation of the working copy includes operations that push to
the monorepo. This was causing a bamboozling error that no such URN
existed. This was in fact true because we were expecting to create the
project after.

This fix places the project creation after the validation but before the
working copy initialisation.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>