rad-identities: improve delegation arguments
It was naive of me to think using serde_json to parse a BTreeSet of
KeyOrUrn would do the right thing. It doesn't.

The better approach is to use a Vec and try parse an Urn or a PeerId,
failing if neither works. This lead to the realisation that PeerId would
be a better option for parsing in both indirect and direct delegations.
This is because `rad profile peer` prints out the profile's PeerId so
it's easy to copy and paste.

The fact that we use a Vec for indirect delegations means we must
convert that Vec to a BTreeSet later, which concludes the other changes.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>