docs: update contributing guide
Since we are continuing to use the mailing list for submitting and
discussing patches, we also update our CONTRIBUTING.md guide to include
how to submit a patch series to the mailing list. On top of this, we
provide some scripts to make it easy to take the repetitiveness of
creating and pushing tags.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>