deps: bump radicle-keystore, ed25519-zebra, and rand
ed25519-zebra has updated to 3.0.0[0]. radicle-keystore was bumped to
use this version, which also required rand to be updated too.

This patch updates radicle-keystore as well as ed25519-zebra and rand, in
the same manner. This means we don't pin the dependency of ed25519-zebra
to the git repository. Which means we can also remove it from the
deny.toml entry.

[0]: https://github.com/ZcashFoundation/ed25519-zebra/issues/45#issuecomment-941279948

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>