librad, link-identities, rad-identities: remove Role type
To avoid the prescription of naming (which is hard), we remove the Role
type. Instead of this, when a Peer is denoted as Replicated we simply
have a single generic parameter.

In librad::git::identities::relations, we introduce a new type Persona
which consists of a Person, a flag saying whether the Persona is a
delegate of the original identity, and the `Refs` they may have for the
original identity.

Note that this is enough information to calculate the information we
used before:
* Maintainer means that the flag is set
* Contributor means that the `Refs` is set to `Some`
* Tracker means that neither were set

Changing this type required the `tracked` functions in rad-identities to
also be updated. We introduced a specialised display type for Persona
which unpacks the urn and payload from the Person and keeps the other
two fields.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>