link-async: new crate
Move the executor abstractions to a new crate. Also provide timer utils
(timeout, sleep, interval), as timers are in fact very
executor-specific.

In the process, remove all use of `futures-timer` which heretically
spawns an ambient thread. Also ensure only one version of the `rand`
crate is used in the workspace.

Signed-off-by: Kim Altintop <kim@eagain.st>