link-canonical: add ty_name function
When wanting to provide useful error messages for turning a Value type
into a domain type, it prompted the want to say what type the Value had
upon an unsuccessful match.

This patch adds a function that tells us what type the Value has, as
well as the keys when it's an object.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>