link-identities: use base tree
If no delegations were provided for the update of a project, the
resulting tree would not include the set of indirect delegations. This
ends up causing an error, since the project is in a corrupted state:

---
Error: error traversing the identity history
Caused by:
    the path 'delegations' does not exist in the given tree; class=Tree
    (14); code=NotFound (-3)
---

The fix in this patch is to use the existing revision as the base, and
update the delegates if provided.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>