link-git-protocol: move to link-git crate
We make use of only a fraction of the gitoxide suite of crates, and
where we do, the high-level `git-repository` crate doesn't meet our
needs. Thus, establish a `link-git` crate which re-exports just the
things from gitoxide crates deemed useful, and implements the custom
functionality we need.

This is part of a series: as a first step, only move `link-git-protocol`
to `link-git`.

Signed-off-by: Kim Altintop <kim@eagain.st>