link-git: shareable odb with pack cache
gitoxide requires to load all packfiles of a repository upfront, and
does not offer a meaningful way to share them across threads.

Major git implementations agree on the strategy employed here: load
index (.idx) files eagerly, and data (.pack) files only when needed (as
determined by a match in the corresponding index). The number of loaded
pack files is bounded, so as to limit resource consumption.

The implementation proposed here follows JGit, but with major
simplifications. The advantage over eventually fair locking (eg.
`parking_lot::RwLock`) is that readers never block. This is deemed to be
crucial for the use case, because our readers are typically independent,
ie. tend to access disjoint sets of .pack files in an unoptimised repo.
Of course, real-worlds measurements will need to be taken in order to
back this claim up.

Signed-off-by: Kim Altintop <kim@eagain.st>