link-git: shareable refdb with packed-refs cache
Loaded `packed-refs` can typically be shared, but determining whether
they need to be reloaded is relatively expensive. Thus load them (if
present) when obtaining a `Refdb` handle, but leave it to the caller to
decide how long to operate on the snapshot.

Also fixes a race found in similar functionality provided by
`git-repository`, which doesn't lock the `packed-refs` while determining
the modification time.

Signed-off-by: Kim Altintop <kim@eagain.st>