librad: generic functions for storage
The functions `reference` and `references` both take domain-specific
types, `Reference<One>` and `Reference<Many>` respectively.
Unfortunately, this limits the use of these functions to looking at
references that are under `refs/namespaces/**`.

This patch changes the inputs to be generic parameters that can be
converted into `RefLike` and `RefspecPattern`. This preserves the
meaning of getting a single reference vs multiple references, while also
allowing the direct use of `RefLike` and `RefspecPattern` -- or anything
that can be converted to them.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>