Upgrade link-git to use latest version of git-ref
`git-ref` was was improved such that users aren't forced anymore to
handle packed-refs thanks to an auto-updated shared cache.

As it's possible to still access the previous (but now renamed)
versions of methods that use packed-refs as parameter. Along with
direct access to the latest cached packed-refs (lazy loaded,
auto-reload on modification), one can get snapshot-like access
by reusing the same packed-refs across multiple calls.

All this is in an effort to get closer to a unified API that
can work similarly in a ref-table implementation.

There is also a 'general' store in the works which can be either
loose refs or a ref-table, but that one seems only relevant once
ref-table is actually implemented.

Signed-off-by: Sebastian Thiel <sebastian.thiel@icloud.com>