link-git: retry packed-refs lock
When opening `packed-refs`, we need to take a file lock in order to
guarantee that our recorded mtime and the actual mtime are the same. We
don't want to fail immediately if a repack is in progress, though, so
install a retry. The timeout is rather generous to account for very
large packed-refs.

Signed-off-by: Kim Altintop <kim@eagain.st>