net: add peer to active set when actively connecting
The membership protocol forgot to add a peer to the active set when the
connection is actively established. This happens when the peer is the
terminal node of a forward-join random walk; the originator of the join
message is not necessarily connected to the local peer, so it needs to
try and establish a connection. Adding the originator to the active set
needs to be deferred until the connection is successfully established,
however.

To fix this, a membership `Connect` tick is mapped to its dedicated
counterpart, instead of conflating it with an unreliable send. The
evaluator loop can then handle the state transitions.

Reported-By: Thomas Scholtes <geigerzaehler@axiom.fm>
Signed-off-by: Kim Altintop <kim@eagain.st>