net: broadcast reply to want
Fixes a long-standing FIXME by thinking about it again: if a peer
receives a `Want` message by means of relaying (ie. not directly from
the originator), it would previously try to connect to the originator.
If that failed, the response would be lost. The implied solution of the
FIXME was: it should instead attempt to connect, but if that fails fall
back to broadcasting.

Upon revisitation, it is concluded that it should only broadcast:
connecting is expensive, and more likely to fail than to succeed. On the
other hand, broadcasting is comparatively cheap and more likely to
actually reach the originator; there should be a path from there to here
through which we received their `Want` in the first place.

Signed-off-by: Kim Altintop <kim@eagain.st>