Always include default categories in Refs::compute
The recent changes to make `Refs::compute` more liberal introduced a
backwards incompatibility. Older implementations _always_ include the
"heads", "notes", and "tags" categories in the computed refs - even if
there are no references in those categories. We changed `Refs::compute`
to be less prescriptive about categories and in the process we changed
the logic so that we _don't_ include any categories which don't have
refs in them. This commit introduces logic to still include the default
git categories even if there are no refs in them.

Signed-off-by: Alex Good <alex@memoryandthought.me>