net: keep track of seen gossip messages
The Epidemic Broadcast Trees[0] protocol assumes that every node in the
network keeps track of all messages it has seen (by applying them to the
local state). This is not true for `link`, as nodes are only interested
in a subset of all data on the network. This is similar to GossipSub[1],
which introduces membership groups to route messages to "subscribed"
nodes. We haven't yet implemented membership groups, but had somewhat
optimistically assumed that at least one node in the network would be
interested in any given messages, due to seed nodes.

Under real world network conditions, this may not be a realistic
assumption, because the seed node could be partitioned from the partial
view which is propagating the message. In this case, a message deemed
uninteresting by all participants would be forwarded indefinitely.

The solution proposed here is to emulate globally-consistent storage by
keeping track of all seen messages probabilistically. That is fairly
easy to reason about, and has the nice property of also providing
another layer of defense against flooding. It requires to keep state,
however, and may not be as effective if nodes are restarting frequently.

An alternative approach would be to equip broadcast messages with a
"ttl" counter, which is decremented on every hop (similar to membership
message). Once the counter reaches zero, the message is discarded. This
would require a network upgrade to become effective, though.

[0]: https://www.gsd.inesc-id.pt/~ler/reports/srds07.pdf
[1]: https://github.com/libp2p/specs/tree/master/pubsub/gossipsub

Signed-off-by: Kim Altintop <kim@eagain.st>