link-canonical: Vec-backed Array
The previous implementation of an Array was backed by a BTreeSet. This
would not actually work and lead to lossy problems. For example, if we
had the following:

---
struct Bar(bool, bool)
---

And we constructed and serialised `Bar(true, true)`, then we would end
up with an array `[true]`, losing the information that there was a
struct of two fields, and both values being true.

Instead, we back the Array with a regular, old Vec -- naturally
preserving the values.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>