test: remove saturation test
This test was introduced as part of 900b6cf6 (replication: Ensure git
stream is closed, 2021-04-26), but it is not clear if it would indeed
catch a regression. It is also rather flaky due to relying on the events
channel for determining when all fetches have been completed; it might
very well be that some fetches never get to run due to contention.

Signed-off-by: Kim Altintop <kim@eagain.st>