link-tracking, librad: no-op deletion case
If a deletion is being peformed, and the PreviousValue policy allows
it, then the action is a no-op and does not persist to the Applied
values, ie. neither updated or rejected.

As a result, the Updated version of Deleted and Untracked do not use
Option, but rather fully realised values.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>