link-tracking: fix cobs representation
The cobs representation was not correct as per the RFC, likely due to
multitasking writing the code and the RFC at the same time :)

The previous implementation allowed a wildcard for all `"cobs"`,
whereas the RFC states that the typename can be a wildcard or specific
cob typename. On top of this, the representation allowed a set of
`Filter`s, however, it should be a single `Filter` with a pattern inside.

These are rectified in this patch as follows. We introduce `TypeName`,
which is either a wildcard or the concrete name of the cob type. This
becomes the key for the map of filters.
We also introduce `Pattern`, which is either a wildcard or a set of
object identifiers.

These two changes required us to rework the surface API. We provide an
`Entry` API, akin to the underlying `BTreeMap`. We also provide some
helpers for more explicit manipulation such as setting the policy or
pattern, or inserting/removing objects.

We also provide `allow_all` and `deny_all`, and a `Default` impl for
convenience of construction.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>