librad: tracking migration function
We did not account for migrating from the old tracking storage -- via
the git config -- to the new tracking storage -- via refs. This patch
introduces the functionality for migrating from v1 to v2. We
reintroduce the v1 code under a v1 module. We add a migration module,
containing a `migrate` function that transitions v1 tracking entries
into v2 tracking entries.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>