rfc: amend batch tracking with fusion rules
We outline a new section in the tracking storage RFC that covers how
track and untrack actions compose with each other, assuming the same
Urn and PeerId.

For each composition pair we give the final output and the reasoning
of why its chosen -- considering the cases of when a ref entry exists
or doesn't.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>