Improve badges’s alt text
For accessibility reasons, the HTML Standard has an entire section
laying out the requirements for alt text [1]. This change makes the
badges’s alt text more closely conform to the HTML Standard.

The standard gives specific scenarios and what should be done in those
scenarios. These two scenarios are the most applicable: [2] and [3].
Unfortunately, we don’t know what the images are going to say, so we
can’t follow [3].

[1]: <https://html.spec.whatwg.org/dev/images.html#alt>
[2]: <https://html.spec.whatwg.org/dev/images.html#a-link-or-button-containing-nothing-but-the-image>
[3]: <https://html.spec.whatwg.org/dev/images.html#text-that-has-been-rendered-to-a-graphic-for-typographical-effect>

Signed-off-by: Jason Yundt <jason@jasonyundt.email>