Clarify naming requirements for Sourcehut CI
If a fork of this repo has a name other than “radicle-link”, then
Sourcehut CI won’t work properly. builds.sr.ht clones both the original
repo and the fork, then proceed to run tests on the original, not the
fork. This change makes it clear that your fork’s name has to be
“radicle-link”.

Signed-off-by: Jason Yundt <jason@jasonyundt.email>