tracking: remove pruning from v1
We missed the fact that `untrack` would attempt to prune any of the
remote branches of the peer that was being untracked. This is not the
behaviour we want for migration since we would like to keep the data
intact. We opt to remove the pruning in the v1 code. We ensure it
works by adding to the migration test checking that a remote reference
remains intact post-migration.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>