Replace Remote<LocalUrl> with git2::remote_anonymous
In `Remote<LocalUrl>::push` and `Remote<LocalUrl>::fetch` we were
creating a temporary local remote in the config of the working copy.
Unfortunately libgit2 does not provide facilities to cleanly remove
these temporary remotes. This commit changes the implementation of
`push` and `fetch` to instead use an anonymous remote with the push and
fetchspecs for that remote set from the local config.

Signed-off-by: Alex Good <alex@memoryandthought.me>