spec: convert to asciidoc
Mostly just some light editing of a pandoc conversion. No longer needed
stuff is either removed or moved to the historical "drafts" folder. The
GH pages action is removed, it is still an open question how to create
those without giving GH push access **and** with some assurance that the
rendered AsciiDoc is legible.

Fixes #704
Signed-off-by: Kim Altintop <kim@eagain.st>