replication: verify all identity updates
Teach the peek phase to verify all identity updates, not just delegates.
If non-delegate verification fails, drop the entire tree of the
corresponding peer id.

Also fix symref'ing `rad/self` refs to top-level identities: this used
to feed a namespaced refname to `Refdb::refname_to_id`, but that method
does not support cross-namespace lookups. Instead, expose the identities
API through the replication `Identities` trait. Note that the top-level
identity does not need to be verified, as it becomes canonical for the
`rad/self` in question.

Lastly, this reveals a bug in ref update code, which set the expected
value for the symref's source unconditionally instead of inspecting the
value looked up previously.

Signed-off-by: Kim Altintop <kim@eagain.st>