Rename rad-* binaries to lnk-*
The rad-* binaries were initially conceived as having a git plumbing
relationship to git porcelain style tools built on top of them by
(amongst others) the alt-clients team. The intention was that extensions
would be published as binaries of the form rad-* on the PATH being
callable directly via rad <extension>. Whilst this does allow for some
flexibility it does still require that there be some agreement between
extension authors as to what common workflows be called and how they
should interact with each other. This structure does not really reflect
the actual relationship between the primary two teams - which is more
akin to `rustc` and `cargo` than it is to git porcelain and git
plumbing. I.e. alt-clients need the low level tools that link provides
but they are trying to build a specific set of workflows for users of
link repositories which really require their own namespace.

As an example of the problems this causes observe the `rad-help`
command in radicle-client-tools [1]. This command hooks into the
extension mechanisms of the `rad` binary to provide a `rad help` command
which guides the user through setting up the git based replication flow
that the alt-clients team is building. This is confusing in a few ways:

1. It conflicts with the built in `help` command which `rad` provides if
   the `rad-help` command is not present on the PATH.
2. It presents a particular workflow which is not the only way one might
   want to interact with link.

This kind of namespacing problem is likely to become more problematic as
we build out the application architecture RFC. For example, alt-clients
already have a `rad sync` command, which does very diffierent things to
what we have informally discussed on the link team. Of course we could
just choose synonyms for all these commands and coexist in the same
namespace, but that seems very confusing for users.

Instead I propose that we rename the `rad` binary. In this patch I've
used the name `lnk` as suggested by Kim. alt-clients would then take
over the name `rad` and making sure that the `lnk` binary is available
for lower level operations becomes a packaging concern.

`lnk` appears to be relatively free of conflicts. There are a number of
tools which have `lnk` in the name for interacting with windows
shortcut files but I believe that shouldn't be a problem.

As well as renaming the binaries and crates we also move the cli crates
into a folder called `cli` to mitigate the confusing similarities
between e.g. link-identities and lnk-identities.

[1]: https://github.com/radicle-dev/radicle-client-tools

Signed-off-by: Alex Good <alex@memoryandthought.me>