link-replication: enforce sigrefs
Teach the fetch phase to send `want`s based on sigrefs only, instead of
filtering advertised refs again. Instead, use `ls-refs` to find sigrefs
of transitively tracked peers, and run another fetch using those.

This effectively enforces that sigrefs and actually fetched refs are
consistent, and prevents anything unsigned from being fetched. Note that
it also forces the remote end to be able to serve the objects referenced
by the sigrefs it advertises -- if a `want` references a non-existent
object, the entire transaction fails.

As a concession to our alt friends, we allow sigrefs of transitive
trackings to be invalid, in which case they will be skipped. If they're
valid, but refer to non-existent objects, however, the transaction may
still fail.

Also apply identity updates before the fetch phase, the reasoning of
which was outlined in 5794134e (RFC: Mutual Synchronisation, 2022-01-04),
respectively:

    https://lists.sr.ht/~radicle-link/dev/%3C20220106191802.13292-1-kim%40eagain.st%3E#%3C20220118152238.GD3642@schmidt.localdomain%3E

Signed-off-by: Kim Altintop <kim@eagain.st>