bins: add Cargo.lock
The original reason to create a `bins` sub-workspace was so that we
could develop binaries that were pinned using a `Cargo.lock` file
without having to have a completely separate repository. When we did
this, however, we missed the fact that the `.gitignore` file was
causing the `bins/Cargo.lock` file to be ignored and so we missed
checking it in.

In this patch we tell the `.gitignore` file to allow the
`bins/Cargo.lock` file and add the `Cargo.lock` file to the
repository.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>