link-replication: fix ref scan
The actual RefScan impl never worked correctly, as it wasn't setting
the namespace prefix properly.

While at it, peeling (sans odb access) is also rectified: we retain the
original refname and the target (which is necessary for transactional
updates), and follow symrefs to find a peeled oid.

Also acknowledge that we can always convert from `Qualified` to gitoxide
types infallibly (but not the other way round).

Signed-off-by: Kim Altintop <kim@eagain.st>