node-lib: refactor code organisation
This patch takes the opportunity to reorganise some types and modules
in node-lib.

The wire types are split into request and response submodules, that
publicly re-export Request and Response, respectively. Any of the
other types can be accessed via the module, but it's noted that they
are not currently needed because of the From/TryFrom instances.

The Announce type was repeated in many places having the same
structure. This patch places it under api::announce and is reused
where needed.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>